refactor: diplomacy relation fixtures and test consolidation (#3825)
Add diplomacy_relation_fixtures.dart; merge relation-updates, new-relation,
and upsert-index suites; migrate decay and favoured-partner tests to shared
builders.

Refs #3825
