+++ packages/colonizethis_diplomacy/test/diplomacy/favoured_trading_partner_test.dart
@@ -3,6 +3,7 @@ import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../support/diplomacy_game_fixtures.dart';
+import '../support/diplomacy_relation_fixtures.dart';
 
 /// Favoured trading partner lookup (Refs #3753 R7.1): the GP a Minor/Tribe
 /// trades with preferentially — the colony suzerain when colonised, else the
@@ -24,20 +25,15 @@ Game _favouredTradingGame({
       colonyStates: colonyStates,
     );
 
-DiplomacyRelation _rel(String gpId, String targetId, num score) =>
-    DiplomacyRelation(
-      factionId1: gpId,
-      factionId2: targetId,
-      score: score,
-      level: scoreToLevel(score),
-    );
-
 void main() {
   group('favouredTradingPartner (Refs #3753 R7.1)', () {
     test('positive: colony suzerain wins regardless of higher relation', () {
       // gpB holds a higher relation with the Tribe, but gpA is the suzerain.
       final game = _favouredTradingGame(
-        relations: [_rel('gpA', 'tribe1', 40.0), _rel('gpB', 'tribe1', 90.0)],
+        relations: [
+          peaceRelation('gpA', 'tribe1', 40.0),
+          peaceRelation('gpB', 'tribe1', 90.0),
+        ],
         colonyStates: const [
           ColonyState(tribeId: 'tribe1', colonyOfGpId: 'gpA', sinceTurn: 3),
         ],
@@ -47,14 +43,14 @@ void main() {
 
     test('positive: independent Tribe → highest-relation GP (72 vs 68)', () {
       final game = _favouredTradingGame(
-        relations: [_rel('gpA', 'tribe1', 72.0), _rel('gpB', 'tribe1', 68.0)],
+        relations: [peaceRelation('gpA', 'tribe1', 72.0), peaceRelation('gpB', 'tribe1', 68.0)],
       );
       expect(favouredTradingPartner(game, 'tribe1'), 'gpA');
     });
 
     test('positive: Minor Nation → highest-relation GP (72 vs 68)', () {
       final game = _favouredTradingGame(
-        relations: [_rel('gpA', 'minor1', 72.0), _rel('gpB', 'minor1', 68.0)],
+        relations: [peaceRelation('gpA', 'minor1', 72.0), peaceRelation('gpB', 'minor1', 68.0)],
       );
       expect(favouredTradingPartner(game, 'minor1'), 'gpA');
     });
@@ -66,14 +62,14 @@ void main() {
           Player(id: 'gpB', displayName: 'B', isHuman: false),
           Player(id: 'gpA', displayName: 'A', isHuman: false),
         ],
-        relations: [_rel('gpB', 'tribe1', 60.0), _rel('gpA', 'tribe1', 60.0)],
+        relations: [peaceRelation('gpB', 'tribe1', 60.0), peaceRelation('gpA', 'tribe1', 60.0)],
       );
       expect(favouredTradingPartner(game, 'tribe1'), 'gpA');
     });
 
     test('positive: decimal precision decides a near tie (60.1 vs 60.0)', () {
       final game = _favouredTradingGame(
-        relations: [_rel('gpA', 'minor1', 60.0), _rel('gpB', 'minor1', 60.1)],
+        relations: [peaceRelation('gpA', 'minor1', 60.0), peaceRelation('gpB', 'minor1', 60.1)],
       );
       expect(favouredTradingPartner(game, 'minor1'), 'gpB');
     });
@@ -85,7 +81,7 @@ void main() {
 
     test('negative: only unrelated GPs (relations with other targets) → null', () {
       final game = _favouredTradingGame(
-        relations: [_rel('gpA', 'minorOther', 80.0)],
+        relations: [peaceRelation('gpA', 'minorOther', 80.0)],
       );
       expect(favouredTradingPartner(game, 'minor1'), isNull);
     });
@@ -93,7 +89,7 @@ void main() {
     test(
       'negative: a single GP with a relation is the favoured partner',
       () {
-        final game = _favouredTradingGame(relations: [_rel('gpB', 'tribe1', 30.0)]);
+        final game = _favouredTradingGame(relations: [peaceRelation('gpB', 'tribe1', 30.0)]);
         expect(favouredTradingPartner(game, 'tribe1'), 'gpB');
       },
     );
